api/requirements.txt
sanic==20.6.3
pynamodb==4.3.3
PyJWT==1.7.1
